LayoutTests/webrtc/audio-replace-track.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Testing local audio capture playback causes "playing" event to fire</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src ="routines.js"></script>
    <script>
    async function doHumAnalysis(stream, expected)
    {
        for (var cptr = 0; cptr < 10; cptr++) {
            var results = await analyseAudio(stream, 200);
            if (results.heardHum === expected)
                return true;
        }
        return false;
    }

    promise_test((test) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);

        var sender;
        var firsStream;
        var secondStream;
        var remoteStream;
        return navigator.mediaDevices.getUserMedia({ audio: { sampleRate: { exact: 48000 } } }).then((stream) => {
            firstStream = stream;
            return navigator.mediaDevices.getUserMedia({ audio: { sampleRate: { exact: 48000 } } });
        }).then((stream) => {
            secondStream = stream;
            return new Promise((resolve, reject) => {
                createConnections((firstConnection) => {
                    sender = firstConnection.addTrack(firstStream.getAudioTracks()[0], firstStream);
                }, (secondConnection) => {
                    secondConnection.ontrack = (trackEvent) => { resolve(trackEvent.streams[0]); };
                });
                setTimeout(() => reject("Test timed out"), 5000);
            });
        }).then((stream) => {
            remoteStream = stream;
            return doHumAnalysis(remoteStream, true)
        }).then((result) => {
            assert_true(result, "heard hum 1");
            return sender.replaceTrack(secondStream.getAudioTracks()[0], secondStream);
        }).then(() => {
            assert_true(sender.track === secondStream.getAudioTracks()[0]);
            return waitFor(500);
        }).then((results) => {
            return doHumAnalysis(remoteStream, true)
        }).then((results) => {
            assert_true(results, "heard hum 2");
        });
    }, "Replacing audio track from a peer connection");
    </script>
</head>
<body>
</body>
</html>
